requirement.txt
requests
Pillow
python-dotenv
Flask
